% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjDAP.R
\name{adjDAP}
\alias{adjDAP}
\title{Calculate dbh and basal area for multiple stem trees}
\usage{
adjDAP(dados, dap = "dbh", censos = c("01", "02"), sufixo = ".mm")
}
\arguments{
\item{dados}{data frame with trees data.}

\item{dap}{character string with the variable dhb name}

\item{censos}{vector with census code}

\item{sufixo}{any suffix to add to dbh name (see exemple)}
}
\value{
return a new data with basal area and recalculated dbh for each tree.
}
\description{
Calculate dbh and basal area for multiple stem trees
}
\examples{

\dontrun{
dados<- data.frame(sp= rep(c("sp1", "sp2", "sp3"), 4), dbh01.mm= rep(c("110;12;130;14", "140;13;12", "14;13", "14"), 3), dbh02.mm= rep(c("150;16;17;18", "NA", "18", "18"), 3) , alt=seq(10, 50, len=12), date01="10-10-2002", date02="10-10-2005", status01= "A", status02 =sample(c(rep("A", 10), "D","D")), stringsAsFactors = FALSE, tag=paste(0, 1:12, sep="") )
adjDAP(dados, dap = "dbh", censos = c("01", "02"), sufixo = ".mm")}
}
\references{
\url{https://opendatakit.org/}
}
\seealso{
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
