% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{censoAudit}
\alias{censoAudit}
\title{Check census data with previous census}
\usage{
censoAudit(dir_exp = getwd(), olddata = "peic09.csv", allsubdir = TRUE)
}
\arguments{
\item{dir_exp}{Data export directory}

\item{base_name}{Files prefix name and directory direction without extension}

\item{form_rep}{Form repeat sections names in order}

\item{data_type}{Long or short version of the data}
}
\value{
'odkexp' returns csv files export from odk collect.
}
\description{
Check census data with previous census
}
\examples{

\dontrun{
censoAudit(censofile = "treeA01.txt", olddata = "/home/aao/Ale2016/AleProjetos/PPPeic/censo2025/openDataKit/peic09.csv")
}

}
\references{
\url{https://opendatakit.org/}
}
\seealso{
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
