% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborsPointPattern.R
\name{countNeighbors}
\alias{countNeighbors}
\title{Count neighbors around}
\usage{
countNeighbors(x, y, xlim = c(0, 1), ylim = c(0, 1), r = 0.1)
}
\arguments{
\item{x, y}{Numeric vectors with 'x' and 'y' coordinates in a cartesian space. They must have the same length.}

\item{xlim, ylim}{Numeric vectors with two values defining the 'x' and 'y' axis limit in the cartesian plane.}

\item{r}{dis.tance for neighborhood definition.}
}
\value{
'countNeighbors' returns a integer vector  with counts the number of neighbors around each point at the maximum distance 'r'
}
\description{
Count number of neighbor for each point for a given r distance neighborhood
}
\examples{
\dontrun{
countNeighbors(x = runif(100, 0, 100), y = runif(100, 0, 100), xlim= c(0,100), ylim = c(0,100), r = 10)
}

}
\references{
Baddeley, A.; Rubak, E; Turner, R. 2016. Spatial Point Patterns: Methodology and Applications with R. CRC Press.

Wiegand, T. & Moloney, K.A. 2014. Handbook of Spatial Point-Pattern Analysis in Ecology. CRC Press.
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
\keyword{analysis}
\keyword{neighborhood}
