% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborsPointPattern.R
\name{edgeData}
\alias{edgeData}
\title{Edge Data Correction}
\usage{
edgeData(
  x,
  y,
  xlim = c(0, 1),
  ylim = c(0, 1),
  rMax = min(c(xlim[2], ylim[2]))/4
)
}
\arguments{
\item{x, y}{Numeric vectors with 'x' and 'y' coordinates in a cartesian space. They must have the same length.}

\item{xlim, ylim}{Numeric vectors with two values defining the 'x' and 'y' axis limit in the cartesian plane.}

\item{rMax}{Maximum distance for neighbors spatial statistics.}
}
\value{
`edgeData` returns a list with tree positions:  
`edgeData` data with border torus correction coordinates
`originalData` original data 'x' and 'y' coordinates
`pointsVF` logical vector with the same length of 'edgeData' indicating the 'TRUE' data coordinates and 'FALSE' border adjustment coordinates.
}
\description{
Adjust border effect using a torus correction connecting vertical and horizontal borders for a continuous spaces
}
\examples{
\dontrun{
edgeData(x = runif(10, 0, 1), y= runif(10, 0, 1), xlim = c(0,1), ylim = c(0,1), rMax = 0.25)
}
}
\references{
Baddeley, A.; Rubak, E; Turner, R. 2016. Spatial Point Patterns: Methodology and Applications with R. CRC Press.
Wiegand, T. & Moloney, K.A. 2014. Handbook of Spatial Point-Pattern Analysis in Ecology. CRC Press.
}
\author{
Alexandre Adalardo de Oliveira \email{ecovirtualpackage@gmail.com}
}
\keyword{border}
\keyword{correction}
\keyword{effect}
