% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{expODK}
\alias{expODK}
\title{Export data from okd collect to a csv files}
\usage{
expODK(
  colDir = getwd(),
  stDir = getwd(),
  expDir = getwd(),
  formId = NULL,
  startDate = NULL,
  endDate = NULL
)
}
\arguments{
\item{colDir}{ODK collect data directory, base name will be used for named the exported files.}

\item{stDir}{ODK storage temporary directory address.}

\item{expDir}{Data export directory.}

\item{formId}{ODK form id, defined in the settings of xls and xml forms.}

\item{startDate}{start date to export data from ODK Collect.}

\item{endtDate}{end date to export data from ODK Collect.}
}
\value{
'odkExp' returns csv files exported from odk collect using odkbriefcase java application. One file is create for each repeat in the form.
}
\description{
Export data from okd collect to a csv files
}
\examples{

\dontrun{
odkExp( colDir = getwd(), stDir = getwd(), expDir = getwd(), form_id = "odkform", file_name = "odkdata")
}
}
\references{
\url{https://opendatakit.org/}
}
\seealso{
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
