% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{joinODK}
\alias{joinODK}
\title{Join odk exported csv files on a single data frame object}
\usage{
joinODK(csvDir = "censoPPSP", saveFile = TRUE, expDir = getwd())
}
\arguments{
\item{csvDir}{directory direction to ODK csv exported files}

\item{saveFile}{logical, if TRUE file will be saved in expDir}

\item{expDir}{Data export directory}
}
\value{
'joinODK' returns txt files that merge ODK csv files exported and keep the csv files in the baseData directory.
}
\description{
Join odk exported csv files on a single data frame object
}
\examples{

\dontrun{
joinODK(csvDir = getwd(),  saveFile = TRUE, expDir = getwd())
}

}
\references{
\url{https://opendatakit.org/}
}
\seealso{
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
