% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborsPointPattern.R
\name{meanNeighbors}
\alias{meanNeighbors}
\title{Mean number of neighbors for different neighborhood sizes}
\usage{
meanNeighbors(
  x,
  y,
  xlim = c(0, 1),
  ylim = c(0, 1),
  rMax = min(c(xlim[2], ylim[2]))/4,
  step = rMax/100
)
}
\arguments{
\item{x, y}{Numeric vectors with 'x' and 'y' coordinates in a cartesian space. They must have the same length.}

\item{xlim, ylim}{Numeric vectors with two values defining the 'x' and 'y' axis limit in the cartesian plane.}

\item{rMax}{maximum distance for neighborhood definition.}

\item{step}{increase distance size for neighborhood.}
}
\value{
'meanNeighbors' returns a data frame with two numeric vectors. The first 'r' represents the neighborhood distance definition; 'meanNeighbors' the mean number of neighbors for each 'r' distance. This count uses the torus border correction.
}
\description{
Calculate mean number of neighbors as a function of distances from each point
}
\examples{
\dontrun{
meanNeighbors(x = runif(100, 0, 1), y = runif(100, 0, 1))
}
}
\references{
Baddeley, A.; Rubak, E; Turner, R. 2016. Spatial Point Patterns: Methodology and Applications with R. CRC Press.
Wiegand, T. & Moloney, K.A. 2014. Handbook of Spatial Point-Pattern Analysis in Ecology. CRC Press.
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
\keyword{analysis}
\keyword{neighborhood}
