% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{mergeData}
\alias{mergeData}
\title{Merge exported data for all subplots}
\usage{
mergeData(
  expDir = getwd(),
  mergeMedia = TRUE,
  saveFile = TRUE,
  zipMedia = FALSE
)
}
\arguments{
\item{expDir}{Data exported directory where the subplots data are arranged in subdirectories.}

\item{mergeMedia}{Logical, if is 'TRUE' média is grouping in 'allMedia' directory but original media was not deleted.}

\item{saveFile}{Lógical, if is 'TRUE' save files in the 'expDir'.}

\item{zipMedia}{Logical if is 'TRUE' the 'allMedia' directory is compressed and deleted. In order to run needs 7-zip installed on the system (\url{https://www.7-zip.org/download.html})}
}
\value{
'mergeData' save two text files('treeRes*.txt' and 'treeAll*.txt') if 'saveFile' is 'TRUE' or/and the resume data in an assign object. .
}
\description{
Merge exported data for all subplots
}
\examples{

\dontrun{
mergeData(expDir = getwd(), mergeMedia = TRUE, zipMedia = FALSE)
}

}
\references{
\url{https://opendatakit.org/}
}
\seealso{
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
