% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborsPointPattern.R
\name{neighbors}
\alias{neighbors}
\title{Identifies neighbors around each point}
\usage{
neighbors(x, y, r)
}
\arguments{
\item{x, y}{Numeric vectors with 'x' and 'y' coordinates in a cartesian space. They must have the same length.}

\item{r}{distance that define neighborhood limits.}

\item{xlim, ylim}{Numeric vectors with two values defining the 'x' and 'y' axis limit in the cartesian plane.}
}
\value{
'neighbors' returns a list of integer vector. Each vector include indexs of neighbors around the point at the maximum distance 'r'. Vector position in the list represents the target point position in the vectors 'x' and 'y'.
}
\description{
Identifies neighbors for each points in a given r distance neighborhood returning the index of neighbor points.
}
\examples{
\dontrun{
neighbors(x = runif(100, 0, 100), y = runif(100, 0, 100), r = 10)
}

}
\references{
Baddeley, A.; Rubak, E; Turner, R. 2016. Spatial Point Patterns: Methodology and Applications with R. CRC Press.
Wiegand, T. & Moloney, K.A. 2014. Handbook of Spatial Point-Pattern Analysis in Ecology. CRC Press.
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
\keyword{analysis}
\keyword{neighborhood}
