% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbioData.R
\name{odb_organize_dataset}
\alias{odb_organize_dataset}
\title{Organize ODBio datasets}
\usage{
odb_organize_dataset(path = getwd())
}
\arguments{
\item{dirPath}{directory address with the uncompressed files exported from openDataBio datasets.}
}
\value{
'odbData' returns a data frame organized as CTFS\\/ForestGEO format.and a
}
\description{
Organize data object from exported opendatabio datasets.

OpenDataBio datasets are exported as zip file containing several tabular data compressed files, with basic information about individuos (Organisms.csv), identification (Taxons.csv) and mapping (Locations.csv). If there is any measurements associated with the individual there is a file (MeasurementsOrFacts.csv)  containing those measurements. This function organize those mesaruments in an format that is used by CTFS/FoestGEO network.
}
\examples{
\dontrun{
dir.create(".Rtmp")
species <- paste("species", 1:10, sep = "_")
family <-  paste("family", 1:10, sep = "_")
spOrder <- sample(1:10, size = 100, replace = TRUE)
x <- sample(seq(0.1, 19.9, by =0.1), size = 100, replace = TRUE)
y <- sample(seq(0.1, 19.9, by =0.1), size = 100, replace = TRUE)
quads <- subplotXY(xcode = LETTERS[1:3], ycode = paste("0", 0:2, sep = "")) 
locationName <- sample(quads$subplot, size =100, replace= TRUE)
ordLoc <- match(locationName, quads$subplot)
gx <- x + quads$xlim[ordLoc] 
gy <- y + quads$ylim[ordLoc]
recordNumber = sample(1:10000, size = 100)

write.table(x = data.frame(id = 1:100,
                           recordNumber,  
                           organismID = paste(recordNumber,"-USER-", locationName, sep = ""),
                           scientificName = species[spOrder],
                           family = family[spOrder],
                           locationName = locationName,
                           locationParentName = "paradisePlot",
                           decimalLatitude =NA,
                           decimalLongitude = NA,
                           x = x,
                           y = y,
                           gx = gx,
                           gy = gy,
                           datasetName = "imaginationPlot"),
            file =".Rtmp/.testOrganisms.csv", sep = ",", row.names = FALSE)
dap <- exp(rnorm(100, mean = 3.68, sd = 0.91))
dap[dap > 10100]<- 64
dap[dap < 10]<- 10
dap <- round(dap)
alt <- exp(-0.73 + 0.63* log(dap) + rnorm(100, 0, 0.35))
alt <- round(alt, 1)

write.table(data.frame(id = 1:200,
                       measured_id = c(1:100, 1:100),
                       measurementType = rep(c("dbh", "height"), each = 100),
                       measurementValue = c(dap, alt),
                       measurementDeterminedDate = "2019-01-10",
                       resourceRelationshipID = paste(recordNumber,"-USER-", locationName, sep = ""),
                       datasetName = "imaginationPlot" ),
            ".Rtmp/.testMeasurementsOrFacts.csv", sep = ",", row.names = FALSE)
dataForestGEO <- odb_organize_dataset(path = ".Rtmp")
str(dataForestGeo)
unlink(""
}
}
\seealso{
\code{\link{data.frame}} 
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
