% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{odkexp}
\alias{odkexp}
\title{Export data from okd collect to a csv files}
\usage{
odkexp(
  odkbriefcase = "odkbriefcase.jar",
  dir_st = getwd(),
  dir_col = getwd(),
  dir_exp = getwd(),
  form_id = NULL,
  file_name = "censoPeic",
  start = NULL,
  end = NULL
)
}
\arguments{
\item{odkbriefcase}{ODK briefcase java jar file.}

\item{dir_st}{ODK storage directory address.}

\item{dir_col}{ODK collect directory.}

\item{dir_exp}{Data export directory}

\item{form_id}{ODK form id, defined in the settings of xls and xml forms}

\item{file_name}{prefix name for csv files.}
}
\value{
'odkexp' returns csv files export from odk collect.
}
\description{
Export data from okd collect to a csv files
}
\examples{

\dontrun{
odkexp(odkbriefcase = "odkbriefcase.jar", dir_st = getwd(), dir_col = getwd(), dir_exp = getwd(), form_id = "odkform", file_name = "odkdata")
}
}
\references{
\url{https://opendatakit.org/}
}
\seealso{
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
