% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{prestaConta}
\alias{prestaConta}
\title{Create html accountability report from prestaConta okd collect form data}
\usage{
prestaConta(
  dataDir,
  dataName = "prestaConta",
  dirExp = getwd(),
  valParc = c(70, 100),
  valFixo = data.frame(nome = c("Renan"), cargo = c("coordena"), valor = c(1500)),
  pagaFixo = FALSE,
  saveFile = TRUE,
  fileRmd = system.file("rmd", "prestaConta.rmd", package = "Rppsp")
)
}
\arguments{
\item{dataDir}{character string, directory where a odk briefcases exports csv files.}

\item{dataName}{character string, odk briefcase data exported prefix file name}

\item{dirExp}{character string, directory where html report and data files will be created.}

\item{valFixo}{data frame with 3 variables, name of field workers,  job role  and  montly paiment.}

\item{pagaFixo}{logic. Should include fix montly payment (valFixo) in the report?}

\item{saveFile}{logic. Should save report data?}

\item{fileRmd}{character string indicating the location of rmd template file.}

\item{valPar}{numeric vector length 2. Value paid for assistant and lider field workers}
}
\value{
'prestaConta' returns html report file and can save data in csv format.
}
\description{
Create html accountability report from prestaConta okd collect form data
}
\examples{

\dontrun{
prestaConta(dataDir = getwd(), dataName="prestaConta", dirExp = getwd(), valParc = c(70,100), valFixo = data.frame(nome= c("Renan"), cargo = c("coordena"),valor = c(1500)), pagaFixo = FALSE,  saveFile=FALSE, fileRmd = system.file("rmd", "prestaConta.rmd", package = "Rppsp"))
}

}
\references{
\url{https://opendatakit.org/}
}
\seealso{
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
