% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{readAudit.csvODK}
\alias{readAudit.csvODK}
\title{Read odk exported audit csv files to a single data frame}
\usage{
readAudit.csvODK(
  base_files,
  save_file = TRUE,
  dir_exp = getwd(),
  file_name = NULL
)
}
\arguments{
\item{dir_exp}{Data export directory}

\item{base_name}{Files prefix name and directory direction without extension}

\item{form_rep}{Form repeat sections names in order}

\item{data_type}{Long or short version of the data}
}
\value{
'odkexp' returns csv files export from odk collect.
}
\description{
Read odk exported audit csv files to a single data frame
}
\examples{

\dontrun{
odkexp(odkbriefcase = "odkbriefcase.jar", dir_st = getwd(), dir_col = getwd(), dir_exp = getwd(), form_id = "odkform", file_name = "odkdata")
}

}
\references{
\url{https://opendatakit.org/}
}
\seealso{
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
