% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat.R
\name{subplot}
\alias{subplot}
\alias{splitPlot}
\alias{subplotXY}
\alias{xyMin}
\title{Functions to calculate subplot index and coordinates}
\usage{
splitPlot(dx, dy, split.x = 5, split.y = split.x, max.x = 20, max.y = max.x)

subplotXY(
  xcode = LETTERS[1:16],
  ycode = c(paste(0, 0:9, sep = ""), 10:15),
  xsub = 20,
  ysub = 20
)

xyMin(subplot_code, xymin = subplotXY())
}
\arguments{
\item{dx}{x coordinate inside subplots}

\item{dy}{y coordinate inside subplots}

\item{max.x}{maximum x coordinate}

\item{max.y}{maximum y coordinate}

\item{gx}{x coordinate for the all plot}

\item{gy}{y coordinate for the all plot}

\item{sub.size}{size of the subplots}

\item{plot.code}{subplot codes}
}
\value{
return a new coordinates gx and gy or plot.code  with basal area and recalculated dbh for each tree.
}
\description{
'splitPlot' create a index for subplots of any dimension
}
\details{
'subplotXY' calculate the x and y coordinates for subplots

'xyMin' calculate x and y coordinates for a vector of suplots
}
\examples{

\dontrun{
splitPlot(dx = runif(10, 0,20), dy = runif(10, 0, 20), split.x = 5)
xcode <-  LETTERS[1:16]
ycode <-  c(paste(0, 0:9, sep = ""), 10:15)
xyMin(subplot_code = paste(sample(xcode, size = 25, replace = TRUE), sample(ycode, size = 25, replace = TRUE), sep = ""))  
}
}
\references{
\url{https://opendatakit.org/}
}
\seealso{
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
