% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connMysql.R
\name{connMysql}
\alias{connMysql}
\title{Connection to PPSP MySQL database}
\usage{
connMysql(dbname = "ppsp")
}
\arguments{
\item{database}{character string. Name of data base.}

\item{user}{character string.}

\item{census}{integer vector. Census years.}
}
\value{
'connMysql' returns a dataframe with data.
}
\description{
Interactive function to connect and query data from remote  duckewiki database
}
\details{
SSH protocol to stablish connection and query database structured. This function should work on any duckewiki database.
}
\examples{

\dontrun{
connMysql(database = "ppsp",  census = NULL)
}

}
\seealso{
\url{{http://labtrop.ib.com.br}; {http://github/adalardo/duckewiki}}
}
\author{
Alexandre Adalardo de Oliveira \email{ecovirtualpackage@gmail.com}
}
\keyword{access}
\keyword{connection,}
\keyword{database}
\keyword{mysql,}
\keyword{remote}
