% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{joinODK}
\alias{joinODK}
\title{Join odk exported csv files on a single data frame object}
\usage{
joinODK(csvDir = "censoPPSP", saveFile = TRUE, expDir = getwd())
}
\arguments{
\item{saveFile}{logical, if TRUE file will be saved in expDir}

\item{expDir}{Data export directory}

\item{baseName}{Files prefix name and directory direction without extension}

\item{formRep}{Form repeat sections names in order}

\item{dataType}{Long or short version of the data}
}
\value{
'read' returns csv files export from odk collect.
}
\description{
Join odk exported csv files on a single data frame object
}
\examples{

\dontrun{
joinODK(baseName = "censoPeic",  saveFile = TRUE, expDir = getwd())
}

}
\references{
\url{https://opendatakit.org/}
}
\seealso{
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
