% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat.R
\name{subplot}
\alias{subplot}
\alias{splitPlot}
\alias{subplotXY}
\alias{xyMin}
\alias{splitPlotXY}
\alias{splitPlotData}
\title{Functions to calculate subplot index, code and coordinates}
\usage{
splitPlot(dx, dy, splitX = 5, splitY = splitX, maxX = 20, maxY = maxX)

subplotXY(
  xcode = LETTERS[1:16],
  ycode = c(paste(0, 0:9, sep = ""), 10:15),
  xsub = 20,
  ysub = 20
)

xyMin(subplot_code, xymin = subplotXY())

splitPlotXY(subplotxy, splitX = 5, splitY = splitX, maxX = 20, maxY = maxX)

splitPlotData(
  censoData,
  subplotCodes,
  splitQuadXY,
  buffer = 2,
  dbh = "dbh",
  gx = "gx",
  gy = "gy",
  status = "status19"
)
}
\arguments{
\item{dx}{x coordinate for subplot}

\item{dy}{y coordinate inside subplots}

\item{splitX}{x size of subquadrats}

\item{splitY}{y size of subquadrats, default equal do 'splitX'}

\item{maxX}{maximum potential value for 'dx'}

\item{maxY}{maximum potential value for 'dy'}

\item{gx}{x coordinate for the all plot}

\item{gy}{y coordinate for the all plot}

\item{sub.size}{size of the subplots}

\item{max.x}{maximum x coordinate}

\item{max.y}{maximum y coordinate}

\item{plot.code}{subplot codes}
}
\value{
return a new coordinates gx and gy or plot code in the format '0x0' representing origin for subplot
}
\description{
'splitPlot' create a code for subquadrats of any dimension
}
\details{
'subplotXY' calculate the x and y coordinates for subplots

'xyMin' calculate x and y coordinates for a vector of suplots

'subplotIndex' calculate x and y coordinates for a vector of suplots
}
\examples{

\dontrun{
splitPlot(dx = runif(10, 0,20), dy = runif(10, 0, 20), split.x = 5)
xcode <-  LETTERS[1:16]
ycode <-  c(paste(0, 0:9, sep = ""), 10:15)
xyMin(subplot_code = paste(sample(xcode, size = 25, replace = TRUE), sample(ycode, size = 25, replace = TRUE), sep = ""))  
}
}
\references{
\url{https://opendatakit.org/}
}
\seealso{
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
