% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsvg.R
\name{svgMap}
\alias{svgMap}
\title{Graphic of trees maped}
\usage{
svgMap(
  mapData,
  subPlotCode = "A00",
  svgSave = TRUE,
  wd2save = file.path(getwd(), subPlotCode),
  dx = "dx",
  dy = "dy",
  tag = "tag",
  dbh = "dbh",
  status = "status",
  mapsize = c(13, 13),
  diagonal = FALSE
)
}
\arguments{
\item{dx}{between 0 to max.size . Tree mapping X coordinate.}

\item{dy}{between 0 to max.size . Tree mapping Y coordinate.}

\item{size.split}{the size of the subquadrat to map in each graphic.}

\item{censo}{data frame from permanent plot tree censo cartezian data.}

\item{quad}{a character string or factor representing some subunit of the permanent plot.}

\item{save.svg}{logical true to save a svg file}

\item{wd}{character string indicating directory towhere the file will be saved.}
}
\value{
'mapsvg' returns tree mapped svg mapped with svg pattern ids.
}
\description{
Map of trees from permanent plot based on cartesian coordinates
}
\examples{
\dontrun{
 svgMap(dataplot, svgSave =FALSE)
}

}
\seealso{
\code{\link{gridSVG}} 
\url{http://labtrop.ib.usp.br}
}
\author{
Alexandre Adalardo de Oliveira \email{aleadalardo@gmail.com}
}
